bug fix with rotation (int..float problem), ui extension remember last window size

+++ dde-ui-extension/src/ua/com/integer/dde/extension/ui/property/imp/common/CommonPropertySupporter.java
@@ -50,7 +50,7 @@ public class CommonPropertySupporter extends PropertySupporter {
 	}
 	
 	private void checkRotation() {
-		actor.setRotation(Integer.parseInt(config.get("actor-rotation", "0")));
+		actor.setRotation(Float.parseFloat(config.get("actor-rotation", "0")));
 	}
 	
 	private void checkScales() {
